pyproject.toml
[project]
name = "orchestrai"
version = "0.2.0"
description = "Multi-agent orchestration control plane for MCP servers"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "langchain>=1.0.0,<2.0.0",
    "langchain-mcp-adapters>=0.2.0",
    "langchain-openai>=0.3.30",
    "langgraph>=0.6.4",
    "mcp>=1.12.4",
    "mcp-use>=1.3.9",
    "python-dotenv>=1.1.1",
    "crewai>=0.95.0",
    "pydantic>=2.7.0",
    "pytest>=8.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["orchestrai"]
